+++ Chrome/assets/options.js
@@ -1,0 +1,1 @@
+window.location = "http://www.furaffinity.net/?ka-options=true";